requirements.txt
pyshorteners==1.0.1
djangorestframework==3.12